pyproject.toml
[tool.poetry]
name = "langtools"
version = "0.1.0"
description = ""
authors = ["なるみ <toucans-cutouts0f@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
openai = "^0.27.8"
langchain = "^0.0.209"
python-dotenv = "^1.0.0"
beautifulsoup4 = "^4.12.2"
pornhub-api = "^0.3.0"
retry = "^0.9.2"
cloudscraper = "^1.2.71"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
yapf = "^0.40.1"
toml = "^0.10.2"
pytest = "^7.3.2"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
